// src/app/models/oficio.model.ts
export interface Oficio {
    id: number;
    ciudadEnvio: string;
    fechaEnvio: Date;
    membrete: string;
    idEntidad: number;
    ciudad: string;
    asunto: string;
    mensaje: string;
    idAlertaSeguimiento: number;
    comentario: string;
    idNNA: number;
    cierre: string;
    firma: string;
    firmaJpg: string;
    userName: string;
}